Remove uses of extclass from reflection.py
