+++ src/exo/reflection.py
@@ -1,96 +1,5 @@
 from . import query_asts as QAST
 from .LoopIR import LoopIR, T
-from .prelude import *
-
-
-@extclass(QAST.Call)
-def __str__(self):
-    return f"{self.proc}(_)"
-
-
-del __str__
-
-
-@extclass(QAST.Alloc)
-def __str__(self):
-    return f"{self.name} : _"
-
-
-del __str__
-
-
-@extclass(QAST.WriteConfig)
-def __str__(self):
-    return f"{self.config.name()}.{self.field} = {str(self.rhs)}"
-
-
-del __str__
-
-
-@extclass(QAST.Assign)
-@extclass(QAST.Reduce)
-def __str__(self):
-    if len(self.idx) > 0:
-        return f"{self.name}[_] = {str(self.rhs)}"
-    else:
-        return f"{self.name} = {str(self.rhs)}"
-
-
-del __str__
-
-
-@extclass(QAST.For)
-def __str__(self):
-    return f"for {self.name} in par(0, {str(self.hi)}):_"
-
-
-del __str__
-
-
-@extclass(QAST.If)
-def __str__(self):
-    cond = str(self.cond)
-    return f"if {cond}:_"
-
-
-del __str__
-
-
-@extclass(QAST.Read)
-def __str__(self):
-    if len(self.idx) > 0:
-        return f"{self.name}[_]"
-    else:
-        return f"{self.name}"
-
-
-del __str__
-
-
-@extclass(QAST.Const)
-def __str__(self):
-    return f"{self.val}"
-
-
-del __str__
-
-
-@extclass(QAST.USub)
-def __str__(self):
-    return f"-{str(self.arg)}"
-
-
-del __str__
-
-
-@extclass(QAST.BinOp)
-def __str__(self):
-    lhs = str(self.lhs)
-    rhs = str(self.rhs)
-    return f"{lhs} {self.op} {rhs}"
-
-
-del __str__
 
 
 # --------------------------------------------------------------------------- #
